requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
black==20.8b1
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
Django==3.1.3
djangorestframework==3.12.2
idna==2.10
mypy-extensions==0.4.3
pathspec==0.8.1
pytz==2020.4
regex==2020.11.11
requests==2.24.0
soupsieve==2.0.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
